PathToPurpose is my new website.
PathToPurpose is my new and first website which I built by realising what is the time left for me to achieve my goals in my life and also got concerned about what is the time left with my mother to watch me get successful. This website helps my users to calculate what is the time left with them to get successful and what is the time left with their parents to watch them to do so.